vrhelmet/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cleanup==4.0.0
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Faker==0.9.1
importlib-metadata==1.6.0
Markdown==3.2.2
mixer==6.1.3
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
zipp==3.1.0
